✨ feat: implement resume upload and workflow components
Adds resume upload, action components, and features list.
Refactors resume page to use new components.
