+++ app/dashboard/resume/components/index.ts
@@ -1,0 +1,5 @@
+export { Header } from './Header';
+export { ResumeUpload } from './ResumeUpload';
+export { ActionDivider } from './ActionDivider';
+export { ActionButton } from './ActionButton';
+export { FeaturesList } from './FeaturesList';
